Add files via upload
This web parser parses the data of job posting in Canada for Data scientists (and similar) from an employment website named 'monster.com'. The web parser uses the request library to get the content of a web page and Beautiful Soup (a python package to parse HTML or XML documents). It uses find() and findAll() methods to filter the meaningful data(here, job titles and details) from HTML tags and elements.